Solve max pairwise product in elixir using sorting.
this solution can be optimized by checking the inputs as they are parsed to make
the algorithm O(n) where n is the size of the numbers passed in. Currently
the runtime is O(n) to read in the input. O(n logn) to sort(assumption).
Then O(n) to  do the look up since getting to the a specific index in a
List is O(n).

O(n) + O(n logn) + (2 * O(n))

Problem: Given a sequence of non-negative integers a0,…,an−1, find the
maximum pairwise product, that is, the largest integer that can be
obtained by multiplying two different elements from the sequence

Input Format: The first line of the input contains an integer n. The
next line contains n non-negative integers a0,...,an−1 (separated by
spaces).

Constraints:
  2 ≤ n ≤ 2 * 10^5
  0 ≤ a0,...,an−1 ≤ 10^5

Output format: Output a single number — the maximum pairwise product.